+++ cuda_modules/matrix_mult_all.cu
@@ -14,7 +14,7 @@ extern "C" {
 #endif
 
 // CUDA Kernel
-__global__ void matrixMultAll(float* C, float* A, float* B, int width, int resW, int resH, int resultSize)
+__global__ void matrixMultAll(double* C, double* A, double* B, int width, int resW, int resH, int resultSize)
 {
 	int x = threadIdx.x + (blockIdx.x * resW);
         int y = threadIdx.y + (blockIdx.y * resH);
