+++ matrix_test.go
@@ -7,20 +7,20 @@ import (
 func TestMultTrans(t *testing.T) {
 	height := 35
 	width := 20
-	m1 := make([][]float32, height)
+	m1 := make([][]float64, height)
 	for i := 0; i < height; i++ {
-		m1[i] = make([]float32, width)
+		m1[i] = make([]float64, width)
 		for j := 0; j < width; j++ {
-			m1[i][j] = float32(i + j)
+			m1[i][j] = float64(i + j)
 		}
 	}
 	height = 35
 	width = 20
-	m2 := make([][]float32, height)
+	m2 := make([][]float64, height)
 	for i := 0; i < height; i++ {
-		m2[i] = make([]float32, width)
+		m2[i] = make([]float64, width)
 		for j := 0; j < width; j++ {
-			m2[i][j] = float32(i + j)
+			m2[i][j] = float64(i + j)
 		}
 	}
 
@@ -39,11 +39,11 @@ func TestMultTrans(t *testing.T) {
 func TestMatrixBigSumAll(t *testing.T) {
 	height := 2500
 	width := 3
-	m1 := make([][]float32, height)
+	m1 := make([][]float64, height)
 	for i := 0; i < height; i++ {
-		m1[i] = make([]float32, width)
+		m1[i] = make([]float64, width)
 		for j := 0; j < width; j++ {
-			m1[i][j] = float32(i + j)
+			m1[i][j] = float64(i + j)
 		}
 	}
 
@@ -56,19 +56,19 @@ func TestMatrixBigSumAll(t *testing.T) {
 }
 
 func TestCudaTrans(t *testing.T) {
-	m1 := GetCudaMatrix([][]float32{
-		[]float32{1, 5, 5, 6},
-		[]float32{1, 9, 8, 3},
-		[]float32{-2, 3.5, 2, 4},
+	m1 := GetCudaMatrix([][]float64{
+		[]float64{1, 5, 5, 6},
+		[]float64{1, 9, 8, 3},
+		[]float64{-2, 3.5, 2, 4},
 	})
 
 	result := m1.Trans().GetMatrixFromCuda()
 
-	expectedRes := [][]float32{
-		[]float32{1, 1, -2},
-		[]float32{5, 9, 3.5},
-		[]float32{5, 8, 2},
-		[]float32{6, 3, 4},
+	expectedRes := [][]float64{
+		[]float64{1, 1, -2},
+		[]float64{5, 9, 3.5},
+		[]float64{5, 8, 2},
+		[]float64{6, 3, 4},
 	}
 
 	for i := 0; i < len(result); i++ {
@@ -81,18 +81,18 @@ func TestCudaTrans(t *testing.T) {
 }
 
 func TestCudaAddBias(t *testing.T) {
-	m1 := GetCudaMatrix([][]float32{
-		[]float32{1, 5},
-		[]float32{1, 9},
-		[]float32{-2, 3.5},
+	m1 := GetCudaMatrix([][]float64{
+		[]float64{1, 5},
+		[]float64{1, 9},
+		[]float64{-2, 3.5},
 	})
 
 	result := m1.AddBias().GetMatrixFromCuda()
 
-	expectedRes := [][]float32{
-		[]float32{1, 1, 5},
-		[]float32{1, 1, 9},
-		[]float32{1, -2, 3.5},
+	expectedRes := [][]float64{
+		[]float64{1, 1, 5},
+		[]float64{1, 1, 9},
+		[]float64{1, -2, 3.5},
 	}
 
 	for i := 0; i < len(result); i++ {
@@ -105,10 +105,10 @@ func TestCudaAddBias(t *testing.T) {
 
 	result = m1.AddBias().RemoveBias().GetMatrixFromCuda()
 
-	expectedRes = [][]float32{
-		[]float32{1, 5},
-		[]float32{1, 9},
-		[]float32{-2, 3.5},
+	expectedRes = [][]float64{
+		[]float64{1, 5},
+		[]float64{1, 9},
+		[]float64{-2, 3.5},
 	}
 
 	for i := 0; i < len(result); i++ {
@@ -121,18 +121,18 @@ func TestCudaAddBias(t *testing.T) {
 }
 
 func TestCudaNegMatrix(t *testing.T) {
-	m1 := [][]float32{
-		[]float32{1, 5},
-		[]float32{1, 9},
-		[]float32{-2, 3.5},
+	m1 := [][]float64{
+		[]float64{1, 5},
+		[]float64{1, 9},
+		[]float64{-2, 3.5},
 	}
 
 	result := Neg(m1)
 
-	expectedRes := [][]float32{
-		[]float32{-1, -5},
-		[]float32{-1, -9},
-		[]float32{2, -3.5},
+	expectedRes := [][]float64{
+		[]float64{-1, -5},
+		[]float64{-1, -9},
+		[]float64{2, -3.5},
 	}
 
 	for i := 0; i < len(result); i++ {
@@ -146,23 +146,23 @@ func TestCudaNegMatrix(t *testing.T) {
 
 func TestCudaNoCudaMult(t *testing.T) {
 
-	m1 := [][]float32{
-		[]float32{1, 5},
-		[]float32{1, 9},
-		[]float32{1, 6},
-		[]float32{1, 7},
-		[]float32{1, 8},
-		[]float32{1, 8},
-		[]float32{1, 5},
-		[]float32{1, 4},
-		[]float32{1, 1},
-		[]float32{1, 4},
-		[]float32{1, 7},
-		[]float32{1, 2},
+	m1 := [][]float64{
+		[]float64{1, 5},
+		[]float64{1, 9},
+		[]float64{1, 6},
+		[]float64{1, 7},
+		[]float64{1, 8},
+		[]float64{1, 8},
+		[]float64{1, 5},
+		[]float64{1, 4},
+		[]float64{1, 1},
+		[]float64{1, 4},
+		[]float64{1, 7},
+		[]float64{1, 2},
 	}
-	m2 := [][]float32{
-		[]float32{1},
-		[]float32{3},
+	m2 := [][]float64{
+		[]float64{1},
+		[]float64{3},
 	}
 
 	cudaRes := Mult(m1, m2)
@@ -179,18 +179,18 @@ func TestCudaNoCudaMult(t *testing.T) {
 func TestMatrixBigMultiplication(t *testing.T) {
 	height := 40
 	width := 60
-	m1 := make([][]float32, height)
+	m1 := make([][]float64, height)
 	for i := 0; i < height; i++ {
-		m1[i] = make([]float32, width)
+		m1[i] = make([]float64, width)
 		for j := 0; j < width; j++ {
-			m1[i][j] = float32(i + j)
+			m1[i][j] = float64(i + j)
 		}
 	}
-	m2 := make([][]float32, width)
+	m2 := make([][]float64, width)
 	for i := 0; i < width; i++ {
-		m2[i] = make([]float32, height)
+		m2[i] = make([]float64, height)
 		for j := 0; j < height; j++ {
-			m2[i][j] = float32(i + j)
+			m2[i][j] = float64(i + j)
 		}
 	}
 
@@ -209,18 +209,18 @@ func TestMatrixBigMultiplication(t *testing.T) {
 func TestMatrixBigSub(t *testing.T) {
 	height := 40
 	width := 60
-	m1 := make([][]float32, height)
+	m1 := make([][]float64, height)
 	for i := 0; i < height; i++ {
-		m1[i] = make([]float32, width)
+		m1[i] = make([]float64, width)
 		for j := 0; j < width; j++ {
-			m1[i][j] = float32(i + j)
+			m1[i][j] = float64(i + j)
 		}
 	}
-	m2 := make([][]float32, height)
+	m2 := make([][]float64, height)
 	for i := 0; i < height; i++ {
-		m2[i] = make([]float32, width)
+		m2[i] = make([]float64, width)
 		for j := 0; j < width; j++ {
-			m2[i][j] = float32(i + j) + 0.5
+			m2[i][j] = float64(i + j) + 0.5
 		}
 	}
 
@@ -237,19 +237,19 @@ func TestMatrixBigSub(t *testing.T) {
 }
 
 func TestCudaMatrixMultiplication(t *testing.T) {
-	m1 := GetCudaMatrix([][]float32{
-		[]float32{3, 2, 1},
-		[]float32{9, 5, 7},
+	m1 := GetCudaMatrix([][]float64{
+		[]float64{3, 2, 1},
+		[]float64{9, 5, 7},
 	})
-	m2 := GetCudaMatrix([][]float32{
-		[]float32{2, 3},
-		[]float32{1, 4},
-		[]float32{2, 9},
+	m2 := GetCudaMatrix([][]float64{
+		[]float64{2, 3},
+		[]float64{1, 4},
+		[]float64{2, 9},
 	})
 
-	expectedRes := [][]float32{
-		[]float32{10, 26},
-		[]float32{37, 110},
+	expectedRes := [][]float64{
+		[]float64{10, 26},
+		[]float64{37, 110},
 	}
 
 	for i := 0; i < 10000; i++ {
@@ -266,19 +266,19 @@ func TestCudaMatrixMultiplication(t *testing.T) {
 }
 
 func TestMatrixMultiplication(t *testing.T) {
-	m1 := [][]float32{
-		[]float32{3, 2, 1},
-		[]float32{9, 5, 7},
+	m1 := [][]float64{
+		[]float64{3, 2, 1},
+		[]float64{9, 5, 7},
 	}
-	m2 := [][]float32{
-		[]float32{2, 3},
-		[]float32{1, 4},
-		[]float32{2, 9},
+	m2 := [][]float64{
+		[]float64{2, 3},
+		[]float64{1, 4},
+		[]float64{2, 9},
 	}
 
-	expectedRes := [][]float32{
-		[]float32{10, 26},
-		[]float32{37, 110},
+	expectedRes := [][]float64{
+		[]float64{10, 26},
+		[]float64{37, 110},
 	}
 
 	for i := 0; i < 10000; i++ {
@@ -295,18 +295,18 @@ func TestMatrixMultiplication(t *testing.T) {
 }
 
 func TestMatrixSum(t *testing.T) {
-	m1 := [][]float32{
-		[]float32{3, 2, 1},
-		[]float32{9, 5, 7},
+	m1 := [][]float64{
+		[]float64{3, 2, 1},
+		[]float64{9, 5, 7},
 	}
-	m2 := [][]float32{
-		[]float32{2, 3, 4},
-		[]float32{1, 4, 7},
+	m2 := [][]float64{
+		[]float64{2, 3, 4},
+		[]float64{1, 4, 7},
 	}
 
-	expectedRes := [][]float32{
-		[]float32{5, 5, 5},
-		[]float32{10, 9, 14},
+	expectedRes := [][]float64{
+		[]float64{5, 5, 5},
+		[]float64{10, 9, 14},
 	}
 
 	result := Sum(m1, m2)
@@ -321,18 +321,18 @@ func TestMatrixSum(t *testing.T) {
 }
 
 func TestMultElems(t *testing.T) {
-	m1 := [][]float32{
-		[]float32{3, 2, 1},
-		[]float32{9, 5, 7},
+	m1 := [][]float64{
+		[]float64{3, 2, 1},
+		[]float64{9, 5, 7},
 	}
-	m2 := [][]float32{
-		[]float32{2, 3, 4},
-		[]float32{1, 4, 7},
+	m2 := [][]float64{
+		[]float64{2, 3, 4},
+		[]float64{1, 4, 7},
 	}
 
-	expectedRes := [][]float32{
-		[]float32{6, 6, 4},
-		[]float32{9, 20, 49},
+	expectedRes := [][]float64{
+		[]float64{6, 6, 4},
+		[]float64{9, 20, 49},
 	}
 
 	result := MultElems(m1, m2)
@@ -347,18 +347,18 @@ func TestMultElems(t *testing.T) {
 }
 
 func TestMatrixSub(t *testing.T) {
-	m1 := [][]float32{
-		[]float32{3, 2, 1},
-		[]float32{9, 5, 7},
+	m1 := [][]float64{
+		[]float64{3, 2, 1},
+		[]float64{9, 5, 7},
 	}
-	m2 := [][]float32{
-		[]float32{2, 3, 4},
-		[]float32{1, 4, 7},
+	m2 := [][]float64{
+		[]float64{2, 3, 4},
+		[]float64{1, 4, 7},
 	}
 
-	expectedRes := [][]float32{
-		[]float32{1, -1, -3},
-		[]float32{8, 1, 0},
+	expectedRes := [][]float64{
+		[]float64{1, -1, -3},
+		[]float64{8, 1, 0},
 	}
 
 	result := Sub(m1, m2)
@@ -373,15 +373,15 @@ func TestMatrixSub(t *testing.T) {
 }
 
 func TestMatrixTrans(t *testing.T) {
-	m1 := [][]float32{
-		[]float32{3, 2, 1},
-		[]float32{9, 5, 7},
+	m1 := [][]float64{
+		[]float64{3, 2, 1},
+		[]float64{9, 5, 7},
 	}
 
-	expectedRes := [][]float32{
-		[]float32{3, 9},
-		[]float32{2, 5},
-		[]float32{1, 7},
+	expectedRes := [][]float64{
+		[]float64{3, 9},
+		[]float64{2, 5},
+		[]float64{1, 7},
 	}
 
 	result := Trans(m1)
@@ -396,9 +396,9 @@ func TestMatrixTrans(t *testing.T) {
 }
 
 func TestSumAll(t *testing.T) {
-	m := [][]float32{
-		[]float32{3, 2, 1},
-		[]float32{9, 5, 7},
+	m := [][]float64{
+		[]float64{3, 2, 1},
+		[]float64{9, 5, 7},
 	}
 
 	result := SumAll(m)
@@ -409,17 +409,17 @@ func TestSumAll(t *testing.T) {
 }
 
 func TestApply(t *testing.T) {
-	m := [][]float32{
-		[]float32{4, 2, 1},
-		[]float32{8, 3, 6},
+	m := [][]float64{
+		[]float64{4, 2, 1},
+		[]float64{8, 3, 6},
 	}
 
-	expectedRes := [][]float32{
-		[]float32{2, 1, 0.5},
-		[]float32{4, 1.5, 3},
+	expectedRes := [][]float64{
+		[]float64{2, 1, 0.5},
+		[]float64{4, 1.5, 3},
 	}
 
-	f := func(x float32) float32 {
+	f := func(x float64) float64 {
 		return x / 2
 	}
 
@@ -435,10 +435,10 @@ func TestApply(t *testing.T) {
 }
 
 func TestDet(t *testing.T) {
-	m := [][]float32{
-		[]float32{1, 3, 1},
-		[]float32{1, 1, 2},
-		[]float32{2, 3, 4},
+	m := [][]float64{
+		[]float64{1, 3, 1},
+		[]float64{1, 1, 2},
+		[]float64{2, 3, 4},
 	}
 
 	d := Det(m)
@@ -447,9 +447,9 @@ func TestDet(t *testing.T) {
 		t.Error("Expected Det result -1,", d, "obtained")
 	}
 
-	m = [][]float32{
-		[]float32{3, 3},
-		[]float32{5, 2},
+	m = [][]float64{
+		[]float64{3, 3},
+		[]float64{5, 2},
 	}
 
 	d = Det(m)
@@ -458,11 +458,11 @@ func TestDet(t *testing.T) {
 		t.Error("Expected Det result -4,", d, "obtained")
 	}
 
-	m = [][]float32{
-		[]float32{3, 2, 1, 9},
-		[]float32{2, 7, 5, 4},
-		[]float32{3, 2, 9, 7},
-		[]float32{4, 3, 3, 1},
+	m = [][]float64{
+		[]float64{3, 2, 1, 9},
+		[]float64{2, 7, 5, 4},
+		[]float64{3, 2, 9, 7},
+		[]float64{4, 3, 3, 1},
 	}
 
 	d = Det(m)
@@ -473,16 +473,16 @@ func TestDet(t *testing.T) {
 }
 
 func TestMinors(t *testing.T) {
-	m := [][]float32{
-		[]float32{1, 3, 1},
-		[]float32{1, 1, 2},
-		[]float32{2, 3, 4},
+	m := [][]float64{
+		[]float64{1, 3, 1},
+		[]float64{1, 1, 2},
+		[]float64{2, 3, 4},
 	}
 
-	expectedRes := [][]float32{
-		[]float32{-2, 0, 1},
-		[]float32{9, 2, -3},
-		[]float32{5, 1, -2},
+	expectedRes := [][]float64{
+		[]float64{-2, 0, 1},
+		[]float64{9, 2, -3},
+		[]float64{5, 1, -2},
 	}
 
 	result := Minors(m)
@@ -497,16 +497,16 @@ func TestMinors(t *testing.T) {
 }
 
 func TestCofactor(t *testing.T) {
-	m := [][]float32{
-		[]float32{1, 3, 1},
-		[]float32{1, -1, -2},
-		[]float32{2, 3, 4},
+	m := [][]float64{
+		[]float64{1, 3, 1},
+		[]float64{1, -1, -2},
+		[]float64{2, 3, 4},
 	}
 
-	expectedRes := [][]float32{
-		[]float32{1, -3, 1},
-		[]float32{-1, -1, 2},
-		[]float32{2, -3, 4},
+	expectedRes := [][]float64{
+		[]float64{1, -3, 1},
+		[]float64{-1, -1, 2},
+		[]float64{2, -3, 4},
 	}
 
 	result := Cofactors(m)
@@ -521,16 +521,16 @@ func TestCofactor(t *testing.T) {
 }
 
 func TestInv(t *testing.T) {
-	m := [][]float32{
-		[]float32{1, 3, 1},
-		[]float32{1, 1, 2},
-		[]float32{2, 3, 4},
+	m := [][]float64{
+		[]float64{1, 3, 1},
+		[]float64{1, 1, 2},
+		[]float64{2, 3, 4},
 	}
 
-	expectedRes := [][]float32{
-		[]float32{2, 9, -5},
-		[]float32{0, -2, 1},
-		[]float32{-1, -3, 2},
+	expectedRes := [][]float64{
+		[]float64{2, 9, -5},
+		[]float64{0, -2, 1},
+		[]float64{-1, -3, 2},
 	}
 
 	result := Inv(m)
@@ -545,22 +545,22 @@ func TestInv(t *testing.T) {
 }
 
 func TestDiv(t *testing.T) {
-	m1 := [][]float32{
-		[]float32{1, 2, 3},
-		[]float32{3, 2, 1},
-		[]float32{2, 1, 3},
+	m1 := [][]float64{
+		[]float64{1, 2, 3},
+		[]float64{3, 2, 1},
+		[]float64{2, 1, 3},
 	}
 
-	m2 := [][]float32{
-		[]float32{4, 5, 6},
-		[]float32{6, 5, 4},
-		[]float32{4, 6, 5},
+	m2 := [][]float64{
+		[]float64{4, 5, 6},
+		[]float64{6, 5, 4},
+		[]float64{4, 6, 5},
 	}
 
-	expectedRes := [][]float32{
-		[]float32{7.0/10.0, 3.0/10.0, 0},
-		[]float32{-3.0/10.0, 7.0/10.0, 0},
-		[]float32{6.0/5.0, 1.0/5.0, -1},
+	expectedRes := [][]float64{
+		[]float64{7.0/10.0, 3.0/10.0, 0},
+		[]float64{-3.0/10.0, 7.0/10.0, 0},
+		[]float64{6.0/5.0, 1.0/5.0, -1},
 	}
 
 	result := Div(m1, m2)
@@ -575,22 +575,22 @@ func TestDiv(t *testing.T) {
 }
 
 func TestConcat(t *testing.T) {
-	m1 := [][]float32{
-		[]float32{1, 2, 3},
-		[]float32{3, 2, 1},
-		[]float32{2, 1, 3},
+	m1 := [][]float64{
+		[]float64{1, 2, 3},
+		[]float64{3, 2, 1},
+		[]float64{2, 1, 3},
 	}
 
-	m2 := [][]float32{
-		[]float32{5, 6},
-		[]float32{5, 4},
-		[]float32{6, 5},
+	m2 := [][]float64{
+		[]float64{5, 6},
+		[]float64{5, 4},
+		[]float64{6, 5},
 	}
 
-	expectedRes := [][]float32{
-		[]float32{1, 2, 3, 5, 6},
-		[]float32{3, 2, 1, 5, 4},
-		[]float32{2, 1, 3, 6, 5},
+	expectedRes := [][]float64{
+		[]float64{1, 2, 3, 5, 6},
+		[]float64{3, 2, 1, 5, 4},
+		[]float64{2, 1, 3, 6, 5},
 	}
 
 	result := Concat(m1, m2)
